windows_client/requirements.txt
hidapi>=0.11.0
keyboard>=0.13.5 